requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch --index-url https://download.pytorch.org/whl/cpu
openai-whisper
yt-dlp
python-telegram-bot
requests
